refactor: SecurityUtil 위치 변경
