+++ service/user/src/main/java/jabaclass/user/auth/application/exception/AuthErrorCode.java
@@ -10,9 +10,7 @@ public enum AuthErrorCode implements ErrorCode {
 	INVALID_PASSWORD(HttpStatus.UNAUTHORIZED, "비밀번호가 틀렸습니다."),
 	INVALID_REFRESH_TOKEN(HttpStatus.UNAUTHORIZED, "유효하지 않은 refresh token입니다."),
 	REFRESH_TOKEN_MISMATCH(HttpStatus.UNAUTHORIZED, "refresh token이 일치하지 않습니다."),
-	ALREADY_LOGGED_OUT(HttpStatus.UNAUTHORIZED, "이미 로그아웃된 유저입니다."),
-	AUTHENTICATION_REQUIRED(HttpStatus.UNAUTHORIZED, "인증이 필요합니다."),
-	INVALID_AUTHENTICATION(HttpStatus.UNAUTHORIZED,"인증 정보가 올바르지 않습니다.");
+	ALREADY_LOGGED_OUT(HttpStatus.UNAUTHORIZED, "이미 로그아웃된 유저입니다.");
 
 	private final HttpStatus status;
 	private final String message;
