+++ service/user/src/main/java/jabaclass/user/user/presentation/controller/UserController.java
@@ -1,7 +1,5 @@
 package jabaclass.user.user.presentation.controller;
 
-
-import jabaclass.user.common.util.SecurityUtil;
 import java.util.List;
 import java.util.UUID;
 
@@ -15,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import jabaclass.auth.util.SecurityUtil;
 import jabaclass.user.user.application.usercase.UserUseCase;
 import jabaclass.user.user.presentation.dto.request.ChangeMyEmailRequestDto;
 import jabaclass.user.user.presentation.dto.request.EmailCheckRequestDto;
